libaudiohal: Decrease mLastReplyLifeTimeNs to update the reply more often
Currently the expire time is the minimum of the double of buffer size,
because regular transfers should update the reply more often,
so this is a fallback—and 100 ms in case the buffer is very large

AV sync in Nuplayer needs corresponding time even for cached
position, so sometimes timeNs can not be updated because of big
mLastReplyLifeTimeNs, that will cause audio timestamp can not be updated
timely.

Decrease mLastReplyLifeTimeNs and remove the double of buffer size
to update the reply more often.

Bug: 346698514
Test: AV sync test.

Change-Id: I66cfc4ae53cff0864141dae554c903ce10c8b87e
(cherry picked from 8882b4f9df720ee4542f6da5a414ab0da8079d59)