Add SourceMetadata as a new parameter in openOutputStream
In order to bypass an audio-encoder and achieve low latency, we need to
pass the USAGE_GAME audio attribute to the audio HAL so that the vendor
can use this audio usage to bypass ms12. The USAGE_GAME will be stored
in SourceMetadata and get pass down to the HAL layer.

Bug: 297875432
Test: N/A
Flag: EXEMPT bugfix
Change-Id: I528cb39ec8f79b672acdbbbf4316ed655f6797ae