audioflinger: initialize strategy to PRODUCT_STRATEGY_NONE
Properly initialize strategy to PRODUCT_STRATEGY_NONE.
In case, when there is no track created for a session, effect is not created at Threads.cpp, instead put in orphan chain. However, at the same time, effect is registered to audiopolicy.

AudioPolicy based on the strategy decides whether it is a music effect or not. As currently, default value is not set, strategy default is set to 0, which does not match to PRODUCT_STRATEGY_NONE, hence effects are not categorized properly as "music" effects in APM, which leads to starting a direct/offload track even though s/w effect is enabled.

Test: Enable s/w only effect, play a track, check it should not use offload/ direct output.

Test: run cts -m android.media.player.cts.MediaPlayerTest#testGapless1

Bug: 350667379

Change-Id: Id8305a7b6ac6ecddf00f6a9b5cb3bb7906857741