seccomp: Fix arm policy for mediaswcodec
The swcodec security policies for arm/arm64 do not include the
crash_dump policy, so they both have to be updated by hand.
Since setsockopt was added to crash_dump and also added to swcodec
policies for arm64 but was missing for arm(32 bit) due to which
crash_dump was not dumping the stack trace for arm. So added
setsockopt for swcodec security policies for arm.

Bug: 349361043
Test: manual
Test: test on Pixel 7 with 32 bit mediaswcodec and see stack
Test: trace for an added crash
Change-Id: Ia1c712b907457951e39ae8953bdd94423e0c8d81