audio policy: fix voice volume update when BT SCO is active
When Bluetooth SCO volume is updated, we must also update the voice call
volume even if the BT SCO software gain is unchanged. This is needed because
the BT SCO gain is maxed out and does not change with BT SCO volume
index: failing to do so can leave the voice call software gain with the
wrong value when a call starts.

Bug: 330843134
Test: repro steps in bug
Flag: EXEMPT bug fix
Change-Id: I1d5b631e919849cd2d672d35edc84fe0fce56dfd