libaudiohal: Modify logging in DeviceHalAidl/Hal2AidlMapper/StreamHalAidl
DeviceHalAidl API can be used for different modules. Modify the logging pattern to include the module names in DeviceHalAidl/Hal2AidlMapper for better information.

Use function name, instance name information embedded in a macro to maintain consistency in the logs.

Similarly, for Stream classes use ioHandle to identify the stream in use.

Test: run CoreAudioHalAidlTest

Bug: 356310696

Change-Id: I4d96408fb984d37f4ac609a9ee3a6adf8411bd89