resourcemanager: update clientInfo when create DeathNotifier
In ResourceManagerService::addResource(), if Pid and Uid are unTrusted,
the mClientInfo stored in DeathNotifier is not updated. At this time,
if the process where Pid is located is killed (lowmemorykiller),
it triggers:
DeathNotifier::binderDied() -> ResourceManagerService::removeResource()
It cannot correctly find pid from mMap(which stores callingpid & callinguid),
so noteStopVideo() will not be called, resulting in the video resource
not being released.

Change-Id: Id456e0e64e34b558ff54507ce010048d9559775b
Signed-off-by: guochuang <guochuang@xiaomi.corp-partner.google.com>